+++ packages/sdk/src/NocturneWalletSDK.ts
@@ -187,6 +187,10 @@ export class NocturneWalletSDK {
     this.logger?.info(
       `Storing optimistic record for op ${opDigest}. Op description: ${metadata?.description}`
     );
+    this.logger &&
+      this.logger?.info(
+        `Storing optimistic record for op ${opDigest}. Description: ${metadata?.description}`
+      );
     await this.db.storeOptimisticRecords(
       opDigest,
       {